+++ test/server/api-controller.spec.js
@@ -117,7 +117,6 @@ describe( 'api', function() {
                         instance: instance,
                         instance_id: instanceId,
                         return_url: ret,
-                        allow_multiple: test.multiple === true ? 'true' : test.multiple,
                         defaults: test.defaults,
                         parent_window_origin: test.parentWindowOrigin
                     } )
@@ -578,7 +577,7 @@ describe( 'api', function() {
                 status: 200,
                 res: {
                     property: 'single_url',
-                    expected: /\/single\/::[a-fA-F0-9]{32}\?/
+                    expected: /\/single\/::[A-z0-9]{4}\?/
                 },
                 offline: true
             }, {
@@ -589,7 +588,7 @@ describe( 'api', function() {
                 status: 200,
                 res: {
                     property: 'single_iframe_url',
-                    expected: /\/single\/i\/::[a-fA-F0-9]{32}\?/
+                    expected: /\/single\/i\/::[A-z0-9]{4}\?/
                 },
                 offline: true
             }, {
@@ -600,7 +599,7 @@ describe( 'api', function() {
                 status: 200,
                 res: {
                     property: 'single_url',
-                    expected: /\/single\/::[a-fA-F0-9]{32}\?/
+                    expected: /\/single\/::[A-z0-9]{4}\?/
                 },
                 offline: true
             }, {
@@ -611,19 +610,54 @@ describe( 'api', function() {
                 status: 200,
                 res: {
                     property: 'single_iframe_url',
+                    expected: /\/single\/i\/::[A-z0-9]{4}\?/
+                },
+                offline: true
+            },
+
+            // /single/once
+            {
+                version: version,
+                endpoint: '/survey/single/once',
+                method: 'get',
+                auth: true,
+                status: 200,
+                res: {
+                    property: 'single_once_url',
+                    expected: /\/single\/::[a-fA-F0-9]{32}\?/
+                },
+                offline: true
+            }, {
+                version: version,
+                endpoint: '/survey/single/once/iframe',
+                method: 'get',
+                auth: true,
+                status: 200,
+                res: {
+                    property: 'single_once_iframe_url',
                     expected: /\/single\/i\/::[a-fA-F0-9]{32}\?/
                 },
                 offline: true
             }, {
                 version: version,
-                endpoint: '/survey/single',
+                endpoint: '/survey/single/once',
                 method: 'post',
                 auth: true,
                 status: 200,
-                multiple: true,
                 res: {
-                    property: 'single_url',
-                    expected: /\/single\/::[A-z0-9]{4}\?/
+                    property: 'single_once_url',
+                    expected: /\/single\/::[a-fA-F0-9]{32}\?/
+                },
+                offline: true
+            }, {
+                version: version,
+                endpoint: '/survey/single/once/iframe',
+                method: 'post',
+                auth: true,
+                status: 200,
+                res: {
+                    property: 'single_once_iframe_url',
+                    expected: /\/single\/i\/::[a-fA-F0-9]{32}\?/
                 },
                 offline: true
             },