+++ CHANGELOG.md
@@ -10,10 +10,10 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ##### Fixed
 - IE11 adds rogue namespaces to ordinal attributes.
 
-[Unreleased] 
+[1.34.0] - 2016-09-19 
 ---------------------
 ##### Added
-- Advanced single submission webform views and API endpoints.
+- Advanced and regular single submission webform views and API endpoints.
 
 [1.33.1] - 2016-09-14
 ---------------------